review - use db package, move db config and functions
