+++ README.md
@@ -45,6 +45,8 @@ You will need to create .env files before running the bot: `dev.env` and `prod.e
 | ROLE_VERIFIED             | The role users receive when validated: create a role, and right click it           |
 | DB_PATH                   | The path of the database file: specify where to save the database file             |
 
+>**_NOTE:_** We recommend the `DB_PATH` point to `./app/db/_store/{db_name}.sqlite`. This path is already added to the .gitignore file.
+
 #### **Commands**
 
 | **Command**                   | use                                                 |
