+++ app/db/config.go
@@ -1,0 +1,12 @@
+package db
+
+import "database/sql"
+
+type Database struct {
+	Conn *sql.DB
+	Cfg  DatabaseCfg
+}
+
+type DatabaseCfg struct {
+	DbPath string
+}
